refactor: address comments, move tag declarations to root
