+++ lib/micromark/mdx-component/index.ts
@@ -1,2 +1,1 @@
-export { GENERIC_MDX_COMPONENT_EXCLUDED_TAGS, INLINE_COMPONENT_TAGS } from './excluded-tags';
 export { mdxComponent } from './syntax';
